integration/audio-spectrograms/job-requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile job-requirements.in
#

apache-beam[gcp]==2.22.0  # via -r job-requirements.in
appdirs==1.4.4            # via pooch
attrs==19.3.0             # via pytest
audioread==2.1.8          # via librosa
avro-python3==1.9.2.1     # via apache-beam
cachetools==3.1.1         # via apache-beam, google-auth
certifi==2020.4.5.2       # via requests
cffi==1.14.2              # via soundfile
chardet==3.0.4            # via requests
crcmod==1.7               # via apache-beam
decorator==4.4.2          # via librosa
dill==0.3.1.1             # via apache-beam
docopt==0.6.2             # via hdfs
fastavro==0.23.4          # via apache-beam
fasteners==0.15           # via google-apitools
future==0.18.2            # via apache-beam
google-api-core[grpc,grpcgcp]==1.20.1  # via google-cloud-bigquery, google-cloud-bigtable, google-cloud-core, google-cloud-datastore, google-cloud-dlp, google-cloud-kms, google-cloud-language, google-cloud-pubsub, google-cloud-spanner, google-cloud-videointelligence, google-cloud-vision
google-apitools==0.5.31   # via apache-beam
google-auth==1.17.2       # via google-api-core, google-cloud-bigquery, google-cloud-storage
google-cloud-bigquery==1.24.0  # via apache-beam
google-cloud-bigtable==1.0.0  # via apache-beam
google-cloud-core==1.3.0  # via apache-beam, google-cloud-bigquery, google-cloud-bigtable, google-cloud-datastore, google-cloud-spanner, google-cloud-storage
google-cloud-datastore==1.7.4  # via apache-beam
google-cloud-dlp==0.13.0  # via apache-beam
google-cloud-kms==1.4.0   # via -r job-requirements.in
google-cloud-language==1.3.0  # via apache-beam
google-cloud-pubsub==1.0.2  # via apache-beam
google-cloud-spanner==1.13.0  # via apache-beam
google-cloud-storage==1.29.0  # via -r job-requirements.in
google-cloud-videointelligence==1.13.0  # via apache-beam
google-cloud-vision==0.42.0  # via apache-beam
google-resumable-media==0.5.1  # via google-cloud-bigquery, google-cloud-storage
googleapis-common-protos[grpc]==1.52.0  # via google-api-core, grpc-google-iam-v1
grpc-google-iam-v1==0.12.3  # via google-cloud-bigtable, google-cloud-kms, google-cloud-pubsub, google-cloud-spanner
grpcio-gcp==0.2.2         # via apache-beam, google-api-core
grpcio==1.29.0            # via apache-beam, google-api-core, googleapis-common-protos, grpc-google-iam-v1, grpcio-gcp
hdfs==2.5.8               # via apache-beam
httplib2==0.17.4          # via apache-beam, google-apitools, oauth2client
idna==2.9                 # via requests
joblib==0.16.0            # via librosa, scikit-learn
librosa==0.8.0            # via -r job-requirements.in
llvmlite==0.34.0          # via numba
mock==2.0.0               # via apache-beam
monotonic==1.5            # via fasteners
more-itertools==8.4.0     # via pytest
numba==0.51.0             # via librosa, resampy
numpy==1.18.5             # via apache-beam, librosa, numba, pyarrow, resampy, scikit-learn, scipy
oauth2client==3.0.0       # via apache-beam, google-apitools
packaging==20.4           # via pooch, pytest
pbr==5.4.5                # via mock
pluggy==0.13.1            # via pytest
pooch==1.1.1              # via librosa
protobuf==3.12.2          # via apache-beam, google-api-core, google-cloud-bigquery, googleapis-common-protos
py==1.9.0                 # via pytest
pyarrow==0.17.1           # via apache-beam
pyasn1-modules==0.2.8     # via google-auth, oauth2client
pyasn1==0.4.8             # via oauth2client, pyasn1-modules, rsa
pycparser==2.20           # via cffi
pydot==1.4.1              # via apache-beam
pymongo==3.10.1           # via apache-beam
pyparsing==2.4.7          # via packaging, pydot
pytest==5.4.3             # via -r job-requirements.in
python-dateutil==2.8.1    # via apache-beam
pytz==2020.1              # via apache-beam, fastavro, google-api-core
requests==2.24.0          # via google-api-core, hdfs, pooch
resampy==0.2.2            # via librosa
rsa==4.6                  # via google-auth, oauth2client
scikit-learn==0.23.2      # via librosa
scipy==1.5.2              # via librosa, resampy, scikit-learn
six==1.15.0               # via fasteners, google-api-core, google-apitools, google-auth, google-cloud-bigquery, google-resumable-media, grpcio, hdfs, mock, oauth2client, packaging, protobuf, python-dateutil, resampy
soundfile==0.10.3.post1   # via librosa
threadpoolctl==2.1.0      # via scikit-learn
typing-extensions==3.7.4.2  # via apache-beam
urllib3==1.25.9           # via requests
wcwidth==0.2.5            # via pytest

# The following packages are considered to be unsafe in a requirements file:
# setuptools
